# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.LoginRequest:
    properties:
      email:
        description: |-
          in: body
          required: true
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.Area:
    properties:
      area_coverage:
        type: number
      area_name:
        type: string
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      district_coverage:
        type: number
      district_id:
        type: string
      id:
        description: swagger:model Area
        type: integer
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
  models.Branch:
    properties:
      aashram_area:
        minimum: 0
        type: number
      address:
        maxLength: 500
        type: string
      branch_members:
        items:
          $ref: '#/definitions/models.BranchMember'
        type: array
      children:
        items:
          $ref: '#/definitions/models.Branch'
        type: array
      city:
        $ref: '#/definitions/models.City'
      city_id:
        minimum: 1
        type: integer
      contact_number:
        maxLength: 20
        type: string
      coordinator_name:
        maxLength: 255
        minLength: 2
        type: string
      country:
        $ref: '#/definitions/models.Country'
      country_id:
        minimum: 1
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      daily_end_time:
        type: string
      daily_start_time:
        type: string
      district:
        $ref: '#/definitions/models.District'
      district_id:
        minimum: 1
        type: integer
      email:
        maxLength: 255
        type: string
      established_on:
        type: string
      id:
        type: integer
      infrastructure:
        items:
          $ref: '#/definitions/models.BranchInfrastructure'
        type: array
      name:
        maxLength: 255
        minLength: 2
        type: string
      open_days:
        maxLength: 100
        type: string
      parent:
        $ref: '#/definitions/models.Branch'
      parent_branch_id:
        type: integer
      pincode:
        type: string
      police_station:
        maxLength: 100
        type: string
      post_office:
        maxLength: 100
        type: string
      state:
        $ref: '#/definitions/models.State'
      state_id:
        minimum: 1
        type: integer
      updated_by:
        type: string
      updated_on:
        type: string
    required:
    - contact_number
    - name
    type: object
  models.BranchInfrastructure:
    properties:
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        minimum: 1
        type: integer
      count:
        minimum: 0
        type: integer
      created_by:
        type: string
      created_on:
        type: string
      id:
        type: integer
      type:
        maxLength: 100
        minLength: 2
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    required:
    - branch_id
    - count
    - type
    type: object
  models.BranchMember:
    properties:
      age:
        maximum: 150
        minimum: 0
        type: integer
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        minimum: 1
        type: integer
      branch_role:
        maxLength: 100
        type: string
      created_by:
        type: string
      created_on:
        type: string
      date_of_birth:
        type: string
      date_of_samarpan:
        type: string
      id:
        type: integer
      member_type:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      qualification:
        maxLength: 255
        type: string
      responsibility:
        maxLength: 500
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    required:
    - branch_id
    - member_type
    - name
    type: object
  models.City:
    properties:
      id:
        type: integer
      name:
        type: string
      state_id:
        type: integer
    type: object
  models.Country:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.District:
    properties:
      country_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      state_id:
        type: integer
    type: object
  models.Donation:
    properties:
      amount:
        type: number
      branch:
        $ref: '#/definitions/models.Branch'
      branch_id:
        type: integer
      created_by:
        description: only set on create
        type: string
      created_on:
        type: string
      donation_type:
        type: string
      event:
        allOf:
        - $ref: '#/definitions/models.Event'
        description: Relations
      event_id:
        type: integer
      id:
        type: integer
      kindtype:
        type: string
      updated_by:
        type: string
      updated_on:
        type: string
    type: object
  models.Event:
    properties:
      event_name:
        type: string
      id:
        type: integer
    type: object
...
  models.State:
    properties:
      country_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: vaish@gmail.com
    name: API Support
  description: API documentation for the DJJS Event Reporting Backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DJJS Event Reporting API
  version: "1.0"
paths:
  /api/areas:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Area'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all areas
      tags:
      - Areas
    post:
      consumes:
      - application/json
      parameters:
      - description: Area payload
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/models.Area'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new area
      tags:
      - Areas
  /api/areas/{id}:
    delete:
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an area
      tags:
      - Areas
    put:
      consumes:
      - application/json
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: area
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an area
      tags:
      - Areas
  /api/areas/search:
    get:
      description: Retrieve an area by its name, or all areas if no name is provided.
      parameters:
      - description: Area Name
        in: query
        name: area_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Area'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get areas by name (or all if none provided)
      tags:
      - Areas
...
  /api/donations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Donation'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all donations
      tags:
      - Donations
    post:
      consumes:
      - application/json
      parameters:
      - description: Donation Payload
        in: body
        name: donation
        required: true
        schema:
          $ref: '#/definitions/models.Donation'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new donation
      tags:
      - Donations
  /api/donations/{id}:
    delete:
      parameters:
      - description: Donation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete donation
      tags:
      - Donations
    put:
      consumes:
      - application/json
      parameters:
      - description: Donation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated fields
        in: body
        name: donation
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update donation
      tags:
      - Donations
...
  /api/events/{event_id}/donations:
    get:
      description: Get all donation records for a specific Event ID
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Donation'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Donations by Event ID
      tags:
      - Donations
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user using email and password
      parameters:
      - description: Login credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user and return JWT token
      tags:
      - Auth
  /logout:
    post:
      description: Clears user's token and ends session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - Auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"